requirements.txt
streamlit
chromadb
sentence-transformers
PyMuPDF
pdfplumber
pytesseract
Pillow
python-dotenv
requests
httpx
deep-translator